+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Add `services.SitewitService.set_site_id()` method.
 * Re-order parameters of `models.Account.create()` and
   `services.SitewitService.create()`.
-* `site_id/client_id` parameter made optional for account creation.
+* Make `site_id/client_id` parameter optional for account creation.
 * Make all field-related parameters of account update optional.
 
 ## 0.7.3